# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EduFlow API Documentation",
        "description": "API documentation for the EduFlow pedagogical management solution.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "EduFlow API Server"
        }
    ],
    "paths": {
...
        "/api/student/interests": {
            "post": {
                "tags": [
                    "Interests"
                ],
                "summary": "Select student interests",
                "operationId": "be23b7c090bb79f11d78873989f0828c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "interest_ids"
                                ],
                                "properties": {
                                    "interest_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Interests selected successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "interests": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/{course}/interests": {
            "post": {
                "tags": [
                    "Interests"
                ],
                "summary": "Attach interests to a course",
                "operationId": "a1343c92f42407e86c3b434d2ebedb78",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "interest_ids"
                                ],
                                "properties": {
                                    "interest_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course interests updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "interests": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Only the course teacher can attach interests"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/student/recommended-courses": {
            "get": {
                "tags": [
                    "Interests"
                ],
                "summary": "Get recommended courses based on interests",
                "operationId": "aa363a0daa58099ab18a4e54dfb4c267",
                "responses": {
                    "200": {
                        "description": "List of recommended courses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "recommended_courses": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payment/checkout/{course}": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Create a Stripe checkout session for a course",
                "operationId": "ace7df1c4e2b9a078f53f136843e3e26",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checkout URL retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "checkout_url": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to create checkout session"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payment/success": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Handle successful payment redirect",
                "operationId": "f12b2f770ad8d8b8cf4b2a345b6e4137",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "student_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment processed and student enrolled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "enrollment": {
                                            "type": "object"
                                        },
                                        "group": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to process success"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT Bearer token **_only_**",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}